pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-email-collector"
version = "1.0.0"
description = "키워드 기반 웹사이트 이메일 수집 CLI 도구"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "tldextract>=5.1.0",
    "anyio>=4.4.0",
]

[project.scripts]
webcollect = "web_email_collector.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["web_email_collector"]
